refactor sign api
