+++ shenyu-examples/shenyu-examples-dubbo/shenyu-examples-alibaba-dubbo-service/src/main/resources/spring-dubbo.xml
@@ -26,7 +26,7 @@
 
     <dubbo:application name="test-dubbo-service"/>
 
-    <dubbo:registry address="zookeeper://localhost:2181"/>
+    <dubbo:registry address="${dubbo.zk}"/>
 
     <dubbo:protocol name="dubbo" port="20888"/>
